scripts/tg-mindshare/requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
httpx==0.27.0
